Split default self-repair helper tests
Move the helper-method coverage into focused child modules with shared test
fakes. Rename the stub builder outcome fields and error variant so the test
intent reads as predicates and collapse the repair-requirement cases into one
parameterised table.
